feat(stocks-backend): add alphavantage requests to build Stock domain

+++ services/stocks_backend/src/main/kotlin/com/marcelogontijo/stocks_backend/core/application/controller/stocks/dto/StockTimeSeriesDto.kt
@@ -18,7 +18,7 @@ data class SeriesDataDto(
     val open: Float,
     val high: Float,
     val low: Float,
-    val volume: Float,
+    val volume: Int,
     val close: Float? = null,
 )
