+++ services/stocks_backend/src/main/kotlin/com/marcelogontijo/stocks_backend/core/domain/stock_time_series/StockTimeSeries.kt
@@ -10,6 +10,6 @@ data class StockTimeSeriesData(
     val open: Float,
     val high: Float,
     val low: Float,
-    val volume: Float,
+    val volume: Int,
     val close: Float? = null
 )
